60-day-bootcamp/day-1-5/profile-enricher/app/routers/health.py
from fastapi import APIRouter
import os
from dotenv import load_dotenv
from typing import Dict

load_dotenv()

ENV: str = os.getenv("ENV", "development")

router = APIRouter()

@router.get("/health")
async def health() -> Dict[str, str]:
    return {"status": "ok", "env": ENV}