60-day-bootcamp/day-1-5/profile-enricher/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic==2.11.0
httpx
python-dotenv==1.2.0
